+++ packages/keyring-eth-qr/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add `HdKeyringV2` class implementing KeyringV2 interface ([#398](https://github.com/MetaMask/accounts/pull/398))
+  - Wraps legacy `QrKeyring` to expose accounts via the unified KeyringV2 API and the `KeyringAccount` type.
+
 ## [1.1.0]
 
 ### Added